Moving palettes as a submodule of color
