+++ src/flamegraph/color/palettes.rs
